+++ tests/functional/dat_reductions/reduc_tests.cpp
@@ -5,7 +5,7 @@
 #endif
 
 #include "op_seq.h"
-#include "op_timing2.h"
+#include "op_profile.h"
 
 #include "../utility.h"
 
@@ -45,7 +45,7 @@ void indirect_dat3_inc(float *n0i, float *n1i, const float *er) {
 int main(int argc, char **argv) {
 
   op_init(argc, argv, 2);
-  op_timing2_start("CppReductionTests");
+  op_profile_start("CppReductionTests");
 
   int my_rank = 0;
   int comm_size = 1;
@@ -192,8 +192,8 @@ int main(int argc, char **argv) {
     printf("direct_dat4_inc passed [rank %d]\n", my_rank);
   }
 
-  op_timing2_finish();
-  op_timing2_output();
+  op_profile_end();
+  op_profile_output();
 
   op_exit();