+++ tests/functional/idx_fortran/idx_tests.F90
@@ -42,7 +42,7 @@ program idx_tests_fortran
   real(8) :: expected
 
   call op_init_base(0, 0)
-  call op_timing2_start("FortranIdxTests")
+  call op_profile_start("FortranIdxTests")
 
   call get_rank_and_size(my_rank, comm_size)
 
@@ -157,10 +157,10 @@ program idx_tests_fortran
 
   deallocate(fetched)
 
-  call op_timing2_finish()
+  call op_profile_end()
   
   if (op_is_root() == 1) print *
-    call op_timing2_output()
+    call op_profile_output()
 
   call op_exit()