+++ tests/functional/strides_fortran/stride_tests.F90
@@ -40,7 +40,7 @@ program stride_tests_fortran
   dim2 = 5
 
   call op_init_base(0, 0)
-  call op_timing2_start("FortranStrideTests")
+  call op_profile_start("FortranStrideTests")
 
   call get_rank_and_size(my_rank, comm_size)
 
@@ -106,10 +106,10 @@ program stride_tests_fortran
   end do
   write(*,*) "write5_within_kernel passed"
 
-  call op_timing2_finish()
+  call op_profile_end()
   
   if (op_is_root() == 1) print *
-    call op_timing2_output()
+    call op_profile_output()
   
   call op_exit()